+++ encoder.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"io"
 	"math"
+	"math/bits"
 	"time"
 )
 
@@ -83,9 +84,10 @@ func EncodeUvarint(w io.Writer, u uint64) (err error) {
 }
 
 func UvarintSize(u uint64) int {
-	var buf [10]byte
-	n := binary.PutUvarint(buf[:], u)
-	return n
+	if u == 0 {
+		return 1
+	}
+	return (bits.Len64(u) + 6) / 7
 }
 
 //----------------------------------------